Fixes to scripts using more symbols.

+++ licensed-facility-crosswalk/update.sh
@@ -6,21 +6,35 @@ if [ "$1" = "--help" ] || [ "$1" = "-h" ]; then
     exit 
 fi
 
+if [ "$1" != "" ] && [ "$1" != "--no-fetch" ]; then
+    echo ""
+    echo "Received bad parameter: \"$1\""
+    echo ""
+    echo "usage: bash update.sh [ --no-fetch ]]"
+    echo ""
+    exit
+fi
+
+uuid1="e89100fd-1f1d-4a37-8205-d588aa42e5a1"
+uuid2="63499f73-feba-43f2-a364-7188d5cf7728"
+id="licensed-facility-crosswalk"
+
 if [ "$1" != "--no-fetch" ]; then
 
-    uuid1="e89100fd-1f1d-4a37-8205-d588aa42e5a1"
-    uuid2="63499f73-feba-43f2-a364-7188d5cf7728"
-
-    /bin/rm -f *.zip *.zip.[0-9]*
+    mv -f *.xls *.xlsx *.json sources/ 2>/dev/null
 
     echo "fetching..."
 
-    wget -q "https://data.chhs.ca.gov/dataset/$uuid1/resource/$uuid2/download/licensed-facility-crosswalk-nox9em.zip"
+    wget -q "https://data.chhs.ca.gov/dataset/$uuid1/resource/$uuid2/download/$id-nox9em.zip"
 
     /bin/rm -f *.json *.xls *.xlsx
 
-    unzip licensed-facility-crosswalk-nox9em.zip
+    unzip $id-nox9em.zip
 fi
 
 # do nothing?
+
+/bin/rm sources/*
+
+mv -f *.xls *.xlsx *.json sources/ 2>/dev/null
